finished installation of FTCLib, and removed old files
