+++ TeamCode/build.gradle
@@ -22,10 +22,9 @@ android {
 dependencies {
     implementation project(':FtcRobotController')
     annotationProcessor files('lib/OpModeAnnotationProcessor.jar')
-
     implementation 'org.apache.commons:commons-math3:3.6.1'
-
     implementation 'com.acmerobotics.roadrunner:core:0.5.6'
-//    implementation 'org.ftclib.ftclib:core:2.0.1'
-    implementation 'org.openftc:easyopencv:1.5.2'
+    implementation 'org.ftclib.ftclib:core:2.0.1'
+    implementation 'org.ftclib.ftclib:vision:2.0.1' // vision
+    implementation 'org.ftclib.ftclib:core:2.0.1' // core
 }
